更改所有DAO层，insert/update/delete返回值为integer。从dalgen模板配置开始改起

+++ imagefetch/common-dal/src/main/java/com/bla/imagefetch/common/dal/imagefactory/auto/dao/ServiceConfigDAO.java
@@ -18,33 +18,33 @@ public class ServiceConfigDAO{
     /**
      * desc:不带Id插入表:service_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long insertWithoutID(ServiceConfigDO entity){
+    public Integer insertWithoutID(ServiceConfigDO entity){
         return serviceConfigDOMapper.insertWithoutID(entity);
     }
     /**
      * desc:带Id插入表:service_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long insertWithID(ServiceConfigDO entity){
+    public Integer insertWithID(ServiceConfigDO entity){
         return serviceConfigDOMapper.insertWithID(entity);
     }
     /**
      * desc:根据Id全量更新表:service_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long updateAll(ServiceConfigDO entity){
+    public Integer updateAll(ServiceConfigDO entity){
         return serviceConfigDOMapper.updateAll(entity);
     }
     /**
      * desc:根据Id删除数据:service_config.<br/>
      * @param id id
-     * @return Long
+     * @return Integer
      */
-    public Long deleteById(Integer id){
+    public Integer deleteById(Integer id){
         return serviceConfigDOMapper.deleteById(id);
     }
     /**
@@ -56,7 +56,7 @@ public class ServiceConfigDAO{
         return serviceConfigDOMapper.queryById(id);
     }
     /**
-     * desc:根据普通索引Name获取数据:task_detail.<br/>
+     * desc:根据普通索引Name获取数据:service_config.<br/>
      * @param name name
      * @return ServiceConfigDO
      */
