+++ imagefetch/common-dal/src/main/java/com/bla/imagefetch/common/dal/imagefactory/auto/dao/TaskConfigDAO.java
@@ -18,33 +18,33 @@ public class TaskConfigDAO{
     /**
      * desc:不带Id插入表:task_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long insertWithoutID(TaskConfigDO entity){
+    public Integer insertWithoutID(TaskConfigDO entity){
         return taskConfigDOMapper.insertWithoutID(entity);
     }
     /**
      * desc:带Id插入表:task_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long insertWithID(TaskConfigDO entity){
+    public Integer insertWithID(TaskConfigDO entity){
         return taskConfigDOMapper.insertWithID(entity);
     }
     /**
      * desc:根据Id全量更新表:task_config.<br/>
      * @param entity entity
-     * @return Long
+     * @return Integer
      */
-    public Long updateAll(TaskConfigDO entity){
+    public Integer updateAll(TaskConfigDO entity){
         return taskConfigDOMapper.updateAll(entity);
     }
     /**
      * desc:根据Id删除数据:task_config.<br/>
      * @param id id
-     * @return Long
+     * @return Integer
      */
-    public Long deleteById(Integer id){
+    public Integer deleteById(Integer id){
         return taskConfigDOMapper.deleteById(id);
     }
     /**
@@ -56,7 +56,7 @@ public class TaskConfigDAO{
         return taskConfigDOMapper.queryById(id);
     }
     /**
-     * desc:根据普通索引Name获取数据:task_detail.<br/>
+     * desc:根据普通索引Name获取数据:task_config.<br/>
      * @param name name
      * @return TaskConfigDO
      */
