+++ imagefetch/core-service/src/main/java/com/bla/imagefetch/core/service/repository/TaskConfigRepository.java
@@ -10,9 +10,9 @@ import com.bla.imagefetch.common.dal.imagefactory.auto.dataobject.TaskConfigDO;
  * @date 2020/7/26 23:06
  */
 public interface TaskConfigRepository {
-    Long insert(TaskConfigDO taskConfigDO);
-    Long update(TaskConfigDO taskConfigDO);
-    Long deleteById(Integer id);
+    Integer insert(TaskConfigDO taskConfigDO);
+    Integer update(TaskConfigDO taskConfigDO);
+    Integer deleteById(Integer id);
     TaskConfigDO queryById(Integer id);
     TaskConfigDO queryByName(String name);
 }
